rename directories: websocket, webrtc, simulator
