+++ packages/dht/src/connection/simulator/SimulatorConnection.ts
