+++ packages/dht/src/dht/DhtNode.ts
@@ -38,7 +38,7 @@ import { RecursiveFinder, RecursiveFindResult } from './find/RecursiveFinder'
 import { StoreRpcLocal } from './store/StoreRpcLocal'
 import { PeerDiscovery } from './discovery/PeerDiscovery'
 import { LocalDataStore } from './store/LocalDataStore'
-import { IceServer } from '../connection/WebRTC/WebRtcConnectorRpcLocal'
+import { IceServer } from '../connection/webrtc/WebRtcConnectorRpcLocal'
 import { registerExternalApiRpcMethods } from './registerExternalApiRpcMethods'
 import { ExternalApiRpcRemote } from './ExternalApiRpcRemote'
 import { UUID } from '../helpers/UUID'
