+++ packages/dht/src/exports.ts
@@ -1,13 +1,13 @@
 export { DhtNode, DhtNodeEvents, DhtNodeOptions } from './dht/DhtNode'
 export { ListeningRpcCommunicator } from './transport/ListeningRpcCommunicator'
-export { Simulator, LatencyType } from './connection/Simulator/Simulator'
-export { SimulatorTransport } from './connection/Simulator/SimulatorTransport'
-export { getRandomRegion, getRegionDelayMatrix } from './connection/Simulator/pings'
+export { Simulator, LatencyType } from './connection/simulator/Simulator'
+export { SimulatorTransport } from './connection/simulator/SimulatorTransport'
+export { getRandomRegion, getRegionDelayMatrix } from './connection/simulator/pings'
 export { PeerDescriptor, Message, NodeType, DataEntry } from './proto/packages/dht/protos/DhtRpc'
 export { ITransport } from './transport/ITransport'
 export { ConnectionManager, ConnectionLocker, PortRange, TlsCertificate } from './connection/ConnectionManager'
 export { DhtRpcOptions } from './rpc-protocol/DhtRpcOptions'
 export { Remote } from './dht/contact/Remote'
 export { areEqualPeerDescriptors } from './helpers/peerIdFromPeerDescriptor'
-export { IceServer } from './connection/WebRTC/WebRtcConnectorRpcLocal'
+export { IceServer } from './connection/webrtc/WebRtcConnectorRpcLocal'
 export { DhtCallContext } from './rpc-protocol/DhtCallContext'
