+++ packages/dht/test/integration/DhtNodeExternalAPI.test.ts
@@ -1,5 +1,5 @@
 import { DhtNode } from '../../src/dht/DhtNode'
-import { LatencyType, Simulator } from '../../src/connection/Simulator/Simulator'
+import { LatencyType, Simulator } from '../../src/connection/simulator/Simulator'
 import { createMockConnectionDhtNode } from '../utils/utils'
 import { Any } from '../../src/proto/google/protobuf/any'
 import { PeerDescriptor } from '../../src/proto/packages/dht/protos/DhtRpc'
