+++ packages/dht/test/unit/WebSocketServer.test.ts
@@ -1,4 +1,4 @@
-import { WebSocketServer } from '../../src/connection/WebSocket/WebSocketServer'
+import { WebSocketServer } from '../../src/connection/websocket/WebSocketServer'
 
 describe('WebSocketServer', () => {
