+++ packages/dht/test/unit/webrtcReplaceInternalIpWithExternalIp.test.ts
@@ -1,4 +1,4 @@
-import { replaceInternalIpWithExternalIp } from '../../src/connection/WebRTC/WebRtcConnectorRpcLocal'
+import { replaceInternalIpWithExternalIp } from '../../src/connection/webrtc/WebRtcConnectorRpcLocal'
 
 describe('replaceIpIfCandidateTypeIsHost', () => {
